renaming uri.cpp to URI.cpp
