+++ src/Config.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "KafkaW/KafkaW.h"
-#include "uri.h"
+#include "URI.h"
 #include <atomic>
 #include <chrono>
 #include <memory>
