+++ src/MainOpt.h
@@ -3,7 +3,7 @@
 #include "ConfigParser.h"
 #include "KafkaW/KafkaW.h"
 #include "SchemaRegistry.h"
-#include "uri.h"
+#include "URI.h"
 #include <memory>
 #include <string>
 #include <vector>
