+++ src/URI.cpp
@@ -1,4 +1,4 @@
-#include "uri.h"
+#include "URI.h"
 #include "logger.h"
 #include <iostream>
